refactor(utils): change filename for better readability
