+++ src/helmoro_utils/helmoro_utils/test_helpers.py
@@ -1,4 +1,4 @@
-# helmoro_utils/test_collection.py
+# helmoro_utils/test_helpers.py
 
 import time
 import rclpy
